refactor: ProductCreateRequest -> ProductRequest로 수정
